Clean up myID, correct borders
This now correctly finds the borders of the current team.  It was using a bad myID from before.  The ID is now stored in the HaliteBot instead of the GameMap.
This code is now ready to use the list of border sites to do something useful.
Improved the output for the logger to include a rough ASCII map that shows current player and borders.