Update code with new edge algorithm
This new algorithm processes from the edge of the territory inward.  It allows for an edge piece to call for help and then have that propogate one level in.
Still needs some work but can consistently beat the gen 1 algorithm.
Need to test against the current algo and prevent the "swapping" from taking place.  Once a piece gets help, it cannot move also.