Improve speed of graph and metrics
This adds the use of a blur on the initial game map metric to provide a better direction for evaluating pieces.  Weights the current value 2 and the average value 1.
Improves the graph creation drastically.  Removes the checks on excluded and instead deletes those afterward and does a check in the network to see that the graph contains the node.
Allows squares to consume each other up to 350 str.
Does not allow the movement of 0 str sqaures.
**Still need to go after the attack game which is quite weak.  The graph will be used though so speed was important.**