Remove core squares, fewer attack options
This attempts to better direct the attack by sorting the targets and removing the core squares.  MAX_DISTANCE also increased to allow more movement.
Also added an A-star like heuristic for the get_path which considers the Manhattan distance from one square to another.
Need to go after the total number of squares being moved.