First attempt at complex adaptive system
Have a simple routine in place to find the border squares that need help.  From there the call goes out to the neighbors and they are all processed.  Currently it just goes to the prev square that had need.
Will run it overnight and see what it does against real opponents.